Almost compiling

+++ crates/environ/src/component.rs
@@ -196,7 +196,7 @@ macro_rules! foreach_builtin_component_function {
             #[cfg(feature = "component-model-async")]
             thread_resume_later(vmctx: vmctx, thread_idx: u32) -> bool;
             #[cfg(feature = "component-model-async")]
-            thread_yield_to(vmctx: vmctx, cancellable: u8, thread_idx: u32) -> u64;
+            thread_yield_to(vmctx: vmctx, cancellable: u8, thread_idx: u32) -> u32;
 
             trap(vmctx: vmctx, code: u8) -> bool;
